profiles: change default governor to nightmare and tune all profiles
All profiles will now use nightmare as the default governor. This is good because the governor in past experiences is very reliable and is quite balanced. I've also tuned all of the governor thresholds according to the profile needs, so battery profiles should now be slightly less aggressive and vice-versa.